NFC: Add build rules for NFC opensource vendor module
Presently the build rules exists within common projects
shared with all the tech teams, while the code exists
in NFC handled components which leads to dependencies.

With android upgrade requirement, the build enablement
for tech specific modules should now be part of the
tech team projects instead of the common.

To meet these requirements, NFC specific build properties
are now part of the NFC projects, thus eliminating the
dependencies with the common projects.

Change-Id: I733145c7c8344c80429fa6b8d96cb2334d57fd6a